requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
sqlparse==0.4.2
drf-yasg==1.20.0
djangorestframework==3.13.1
django-silk==4.3.0
psycopg2==2.9.2
djangorestframework-camel-case==1.3.0
Scrapy==2.6.1
gunicorn==20.1.0
python-dateutil==2.8.2
drf-flex-fields==0.9.8